Add profile view & change password

+++ users/deleteDiary.php
@@ -3,16 +3,20 @@
     require_once "./includes/header.php";
 
     // Delete record
-    if(isset($_GET['token'])) {
-        $id=$_GET['token'];
-        $delete_sql = "DELETE FROM diary_table WHERE ID=".$id;
-        if($conn->query($delete_sql)) {
-            header("location:./index.php");
+    if(isset($_SESSION['id'])) {
+        if(isset($_GET['token'])) {
+            $id=$_GET['token'];
+            $delete_sql = "DELETE FROM diary_table WHERE ID=".$id;
+            if($conn->query($delete_sql)) {
+                header("location:./index.php");
+            } else {
+                echo $conn->error;
+            }
         } else {
-            echo $conn->error;
+            header("location:./index.php");
         }
     } else {
-        header("location:./index.php");
+        header("location:./login.php");
     }
 ?>
